% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataganger-package.R
\docType{package}
\name{dataganger-package}
\alias{dataganger}
\alias{dataganger-package}
\title{DataGangeR uses two synthesis engines, chosen automatically by your
objective. Lower-fidelity objectives use an internal marginal engine. For
Model pipeline prototype and Advanced / internal hi-fi objectives, where
preserving relationships between variables matters, DataGangeR uses the
synthpop package (Nowok, Raab & Dibben, 2016). Install it with
\code{install.packages("synthpop")} to enable these objectives at full fidelity.}
\description{
When synthpop is used, please cite: Nowok B, Raab GM, Dibben C (2016).
"synthpop: Bespoke Creation of Synthetic Data in R." \emph{Journal of
Statistical Software}, 74(11), 1-26. doi:10.18637/jss.v074.i11
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lennon-li/dataganger}
  \item Report bugs at \url{https://github.com/lennon-li/dataganger/issues}
}

}
\author{
\strong{Maintainer}: Lennon Li \email{yeli@biostats.ai}

Authors:
\itemize{
  \item Lennon Li \email{yeli@biostats.ai}
}

}
\keyword{internal}
