% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-api.R
\name{generator_risk_report}
\alias{generator_risk_report}
\title{Return the fitting risk report for a frozen generator}
\usage{
generator_risk_report(frozen)
}
\arguments{
\item{frozen}{A \code{dataganger_frozen_generator} handle.}
}
\value{
A \code{dataganger_generator_risk_report} object.
}
\description{
Return the fitting risk report for a frozen generator
}
