% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-batch.R
\name{summary.dataganger_batch}
\alias{summary.dataganger_batch}
\title{Summarize a synthetic development batch}
\usage{
\method{summary}{dataganger_batch}(object, ...)
}
\arguments{
\item{object}{A \code{dataganger_batch} object.}

\item{...}{Unused.}
}
\value{
A \code{summary_dataganger_batch} object.
}
\description{
Summarize a synthetic development batch
}
