% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataganger-package.R
\docType{package}
\name{dataganger-package}
\alias{dataganger}
\alias{dataganger-package}
\title{DataGangeR uses two synthesis engines. By default the engine is chosen
automatically from your objective: the analytics purpose (and any
high-fidelity setting), where preserving relationships between variables
matters, uses the synthpop package (Nowok, Raab & Dibben, 2016); other
objectives use a dependency-free internal marginal engine. In the Shiny app
the engine can also be selected explicitly (auto, internal, or synthpop).
Install synthpop with \code{install.packages("synthpop")} to enable
relationship-preserving synthesis at full fidelity.}
\description{
When synthpop is used, please cite: Nowok B, Raab GM, Dibben C (2016).
"synthpop: Bespoke Creation of Synthetic Data in R." \emph{Journal of
Statistical Software}, 74(11), 1-26. doi:10.18637/jss.v074.i11
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lennon-li/dataganger}
  \item Report bugs at \url{https://github.com/lennon-li/dataganger/issues}
}

}
\author{
\strong{Maintainer}: Lennon Li \email{yeli@biostats.ai}

Authors:
\itemize{
  \item Lennon Li \email{yeli@biostats.ai}
}

}
\keyword{internal}
